source/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2
django-braces==1.14.0
django-cors-headers==3.2.1
django-oauth-toolkit==1.3.2
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
